requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
git-filter-repo==2.45.0
mysqlclient==2.2.4
phonenumbers==8.13.45
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
